requirements.txt
pandas==1.0.3
PyYAML==5.3.1
numpy==1.18.1
sklearn==0.0
matplotlib==3.1.3
seaborn==0.10.0
requests==2.23.0
pytest==5.3.5